requirements.txt
requests~=2.32.3
beautifulsoup4~=4.13.3
python-dotenv~=1.0.1
selenium~=4.28.1
peewee~=3.17.9
fastapi
uvicorn
bcrypt
pyjwt
pydantic[email]

# testing
pytest
pytest-mock
pytest-sugar
pytest-cov